+++ src/test/java/com/intuit/fuzzymatcher/AppConfig.java
@@ -1,0 +1,10 @@
+package com.intuit.fuzzymatcher;
+
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@ComponentScan(basePackages = "com.intuit.fuzzymatcher")
+public class AppConfig {
+
+}