+++ CHANGELOG.md
@@ -8,3 +8,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## 0.0.1 - 2018-11-13
 ### Added
 - Initial Version
+
+## 0.0.2 - 2018-12-21
+### Added
+- Integration with maven central
+### Changed
+- Group id changed com.intuit.fuzzymatcher from com.intuit.fm
+
+## 0.0.3 - 2018-12-21
+### Changed
+- Package structure changed to com.intuit.fuzzymatcher to match group id convention  
+